Add installation logic for Debian & Ubuntu.
This update should make the .deb release work on all Debian based distributions,
starting with Debian Jessie up to the newest Ubuntu. The python modules are
linked to the old python 3.4 from Debian Jessie, which is then linked to the
current python version on the system using a post-installation script. The
library link is kept completely private to RTest and does not influence the rest
of the system.